Fix more -Weverything warnings.

+++ test/lifetime.cpp
@@ -16,11 +16,11 @@
 
 namespace nda {
 
-int lifetime_counter::default_constructs = 0;
-int lifetime_counter::copy_constructs = 0;
-int lifetime_counter::move_constructs = 0;
-int lifetime_counter::copy_assigns = 0;
-int lifetime_counter::move_assigns = 0;
-int lifetime_counter::destructs = 0;
+size_t lifetime_counter::default_constructs = 0;
+size_t lifetime_counter::copy_constructs = 0;
+size_t lifetime_counter::move_constructs = 0;
+size_t lifetime_counter::copy_assigns = 0;
+size_t lifetime_counter::move_assigns = 0;
+size_t lifetime_counter::destructs = 0;
 
 }  // namespace nda
